Reload journal lazily.
The Dropbox update inotify event stream is complex and sometimes
involves deleting and replacing files. With lazy loading, we can log the
Dropbox sync event and reload upon HTTP request when we know the journal
is ready.

This requires turning off the inotify watch when the lazy reload is set.
Otherwise, the lazy reload triggers an inotify event which triggers
another lazy reload. The inotify watch can be restored after the journal
is read.